requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.3
httpx==0.28.1
python-dotenv==1.0.1
python-telegram-bot==21.10
PyYAML==6.0.2
pytest==8.3.5
